Deprecate grudge.eager in favor of grudge.op
