+++ examples/wave/wave-op-mpi.py
@@ -32,21 +32,20 @@ from meshmode.dof_array import thaw
 
 from meshmode.mesh import BTAG_ALL, BTAG_NONE  # noqa
 
-from grudge.eager import EagerDGDiscretization, interior_trace_pair
+from grudge.discretization import DGDiscretizationWithBoundaries
+import grudge.op as op
 from grudge.shortcuts import make_visualizer
-from grudge.symbolic.primitives import TracePair, QTAG_NONE, DOFDesc
+from grudge.symbolic.primitives import TracePair
+from mpi4py import MPI
 
 
 # {{{ wave equation bits
 
 def wave_flux(discr, c, w_tpair):
-    dd = w_tpair.dd
-    dd_quad = dd.with_qtag("vel_prod")
-
     u = w_tpair[0]
     v = w_tpair[1:]
 
-    normal = thaw(u.int.array_context, discr.normal(dd))
+    normal = thaw(u.int.array_context, op.normal(discr, w_tpair.dd))
 
     flux_weak = flat_obj_array(
             np.dot(v.avg, normal),
@@ -54,49 +53,40 @@ def wave_flux(discr, c, w_tpair):
             )
 
     # upwind
+    v_jump = np.dot(normal, v.ext-v.int)
     flux_weak += flat_obj_array(
             0.5*(u.ext-u.int),
-            0.5*normal*np.dot(normal, v.ext-v.int),
+            0.5*normal*v_jump,
             )
 
-    # FIXME this flux is only correct for continuous c
-    dd_allfaces_quad = dd_quad.with_dtag("all_faces")
-    c_quad = discr.project("vol", dd_quad, c)
-    flux_quad = discr.project(dd, dd_quad, flux_weak)
-
-    return discr.project(dd_quad, dd_allfaces_quad, c_quad*flux_quad)
+    return op.project(discr, w_tpair.dd, "all_faces", c*flux_weak)
 
 
 def wave_operator(discr, c, w):
     u = w[0]
     v = w[1:]
 
-    dir_u = discr.project("vol", BTAG_ALL, u)
-    dir_v = discr.project("vol", BTAG_ALL, v)
+    dir_u = op.project(discr, "vol", BTAG_ALL, u)
+    dir_v = op.project(discr, "vol", BTAG_ALL, v)
     dir_bval = flat_obj_array(dir_u, dir_v)
     dir_bc = flat_obj_array(-dir_u, dir_v)
 
-    dd_quad = DOFDesc("vol", "vel_prod")
-    c_quad = discr.project("vol", dd_quad, c)
-    w_quad = discr.project("vol", dd_quad, w)
-    u_quad = w_quad[0]
-    v_quad = w_quad[1:]
-
-    dd_allfaces_quad = DOFDesc("all_faces", "vel_prod")
-
     return (
-            discr.inverse_mass(
+            op.inverse_mass(discr,
                 flat_obj_array(
-                    -discr.weak_div(dd_quad, c_quad*v_quad),
-                    -discr.weak_grad(dd_quad, c_quad*u_quad)
+                    -c*op.weak_div(discr, v),
+                    -c*op.weak_grad(discr, u)
                     )
                 +  # noqa: W504
-                discr.face_mass(
-                    dd_allfaces_quad,
-                    wave_flux(discr, c=c, w_tpair=interior_trace_pair(discr, w))
+                op.face_mass(discr,
+                    wave_flux(discr, c=c, w_tpair=op.interior_trace_pair(discr, w))
                     + wave_flux(discr, c=c, w_tpair=TracePair(
                         BTAG_ALL, interior=dir_bval, exterior=dir_bc))
-                    ))
+                    + sum(
+                        wave_flux(discr, c=c, w_tpair=tpair)
+                        for tpair in op.cross_rank_trace_pairs(discr, w))
+                    )
+                )
                 )
 
 # }}}
@@ -110,16 +100,14 @@ def rk4_step(y, t, h, f):
     return y + h/6*(k1 + 2*k2 + 2*k3 + k4)
 
 
-def bump(actx, discr, t=0, width=0.05, center=None):
-    if center is None:
-        center = np.array([0.2, 0.35, 0.1])
-
-    center = center[:discr.dim]
+def bump(actx, discr, t=0):
+    source_center = np.array([0.2, 0.35, 0.1])[:discr.dim]
+    source_width = 0.05
     source_omega = 3
 
-    nodes = thaw(actx, discr.nodes())
+    nodes = thaw(actx, op.nodes(discr))
     center_dist = flat_obj_array([
-        nodes[i] - center[i]
+        nodes[i] - source_center[i]
         for i in range(discr.dim)
         ])
 
@@ -127,7 +115,7 @@ def bump(actx, discr, t=0, width=0.05, center=None):
         np.cos(source_omega*t)
         * actx.np.exp(
             -np.dot(center_dist, center_dist)
-            / width**2))
+            / source_width**2))
 
 
 def main():
@@ -135,15 +123,37 @@ def main():
     queue = cl.CommandQueue(cl_ctx)
     actx = PyOpenCLArrayContext(queue)
 
+    comm = MPI.COMM_WORLD
+    num_parts = comm.Get_size()
+
+    from meshmode.distributed import MPIMeshDistributor, get_partition_by_pymetis
+    mesh_dist = MPIMeshDistributor(comm)
+
     dim = 2
     nel_1d = 16
-    from meshmode.mesh.generation import generate_regular_rect_mesh
-    mesh = generate_regular_rect_mesh(
-            a=(-0.5,)*dim,
-            b=(0.5,)*dim,
-            n=(nel_1d,)*dim)
+
+    if mesh_dist.is_mananger_rank():
+        from meshmode.mesh.generation import generate_regular_rect_mesh
+        mesh = generate_regular_rect_mesh(
+                a=(-0.5,)*dim,
+                b=(0.5,)*dim,
+                n=(nel_1d,)*dim)
+
+        print("%d elements" % mesh.nelements)
+
+        part_per_element = get_partition_by_pymetis(mesh, num_parts)
+
+        local_mesh = mesh_dist.send_mesh_parts(mesh, part_per_element, num_parts)
+
+        del mesh
+
+    else:
+        local_mesh = mesh_dist.receive_mesh_part()
 
     order = 3
+
+    discr = DGDiscretizationWithBoundaries(actx, local_mesh, order=order,
+                    mpi_communicator=comm)
 
     if dim == 2:
         # no deep meaning here, just a fudge factor
@@ -154,29 +164,15 @@ def main():
     else:
         raise ValueError("don't have a stable time step guesstimate")
 
-    print("%d elements" % mesh.nelements)
-
-    from meshmode.discretization.poly_element import \
-            QuadratureSimplexGroupFactory, \
-            PolynomialWarpAndBlendGroupFactory
-    discr = EagerDGDiscretization(actx, mesh,
-            quad_tag_to_group_factory={
-                QTAG_NONE: PolynomialWarpAndBlendGroupFactory(order),
-                "vel_prod": QuadratureSimplexGroupFactory(3*order),
-                })
-
-    # bounded above by 1
-    c = 0.2 + 0.8*bump(actx, discr, center=np.zeros(3), width=0.5)
-
     fields = flat_obj_array(
-            bump(actx, discr, ),
+            bump(actx, discr),
             [discr.zeros(actx) for i in range(discr.dim)]
             )
 
     vis = make_visualizer(discr, order+3 if dim == 2 else order)
 
     def rhs(t, w):
-        return wave_operator(discr, c=c, w=w)
+        return wave_operator(discr, c=1, w=w)
 
     t = 0
     t_final = 3
@@ -185,10 +181,11 @@ def main():
         fields = rk4_step(fields, t, dt, rhs)
 
         if istep % 10 == 0:
-            print(istep, t, discr.norm(fields[0]))
-            vis.write_vtk_file("fld-wave-eager-var-velocity-%04d.vtu" % istep,
+            print(istep, t, op.norm(discr, fields[0], p=2))
+            vis.write_parallel_vtk_file(
+                    comm,
+                    f"fld-wave-eager-mpi-{{rank:03d}}-{istep:04d}.vtu",
                     [
-                        ("c", c),
                         ("u", fields[0]),
                         ("v", fields[1:]),
                         ])
