+++ examples/wave/wave-op-var-velocity.py
@@ -32,18 +32,22 @@ from meshmode.dof_array import thaw
 
 from meshmode.mesh import BTAG_ALL, BTAG_NONE  # noqa
 
-from grudge.eager import EagerDGDiscretization, interior_trace_pair
+from grudge.discretization import DGDiscretizationWithBoundaries
+import grudge.op as op
 from grudge.shortcuts import make_visualizer
-from grudge.symbolic.primitives import TracePair
+from grudge.symbolic.primitives import TracePair, QTAG_NONE, DOFDesc
 
 
 # {{{ wave equation bits
 
 def wave_flux(discr, c, w_tpair):
+    dd = w_tpair.dd
+    dd_quad = dd.with_qtag("vel_prod")
+
     u = w_tpair[0]
     v = w_tpair[1:]
 
-    normal = thaw(u.int.array_context, discr.normal(w_tpair.dd))
+    normal = thaw(u.int.array_context, op.normal(discr, dd))
 
     flux_weak = flat_obj_array(
             np.dot(v.avg, normal),
@@ -56,27 +60,41 @@ def wave_flux(discr, c, w_tpair):
             0.5*normal*np.dot(normal, v.ext-v.int),
             )
 
-    return discr.project(w_tpair.dd, "all_faces", c*flux_weak)
+    # FIXME this flux is only correct for continuous c
+    dd_allfaces_quad = dd_quad.with_dtag("all_faces")
+    c_quad = op.project(discr, "vol", dd_quad, c)
+    flux_quad = op.project(discr, dd, dd_quad, flux_weak)
+
+    return op.project(discr, dd_quad, dd_allfaces_quad, c_quad*flux_quad)
 
 
 def wave_operator(discr, c, w):
     u = w[0]
     v = w[1:]
 
-    dir_u = discr.project("vol", BTAG_ALL, u)
-    dir_v = discr.project("vol", BTAG_ALL, v)
+    dir_u = op.project(discr, "vol", BTAG_ALL, u)
+    dir_v = op.project(discr, "vol", BTAG_ALL, v)
     dir_bval = flat_obj_array(dir_u, dir_v)
     dir_bc = flat_obj_array(-dir_u, dir_v)
 
+    dd_quad = DOFDesc("vol", "vel_prod")
+    c_quad = op.project(discr, "vol", dd_quad, c)
+    w_quad = op.project(discr, "vol", dd_quad, w)
+    u_quad = w_quad[0]
+    v_quad = w_quad[1:]
+
+    dd_allfaces_quad = DOFDesc("all_faces", "vel_prod")
+
     return (
-            discr.inverse_mass(
+            op.inverse_mass(discr,
                 flat_obj_array(
-                    -c*discr.weak_div(v),
-                    -c*discr.weak_grad(u)
+                    -op.weak_div(discr, dd_quad, c_quad*v_quad),
+                    -op.weak_grad(discr, dd_quad, c_quad*u_quad)
                     )
                 +  # noqa: W504
-                discr.face_mass(
-                    wave_flux(discr, c=c, w_tpair=interior_trace_pair(discr, w))
+                op.face_mass(discr,
+                    dd_allfaces_quad,
+                    wave_flux(discr, c=c, w_tpair=op.interior_trace_pair(discr, w))
                     + wave_flux(discr, c=c, w_tpair=TracePair(
                         BTAG_ALL, interior=dir_bval, exterior=dir_bc))
                     ))
@@ -93,14 +111,16 @@ def rk4_step(y, t, h, f):
     return y + h/6*(k1 + 2*k2 + 2*k3 + k4)
 
 
-def bump(actx, discr, t=0):
-    source_center = np.array([0.2, 0.35, 0.1])[:discr.dim]
-    source_width = 0.05
+def bump(actx, discr, t=0, width=0.05, center=None):
+    if center is None:
+        center = np.array([0.2, 0.35, 0.1])
+
+    center = center[:discr.dim]
     source_omega = 3
 
-    nodes = thaw(actx, discr.nodes())
+    nodes = thaw(actx, op.nodes(discr))
     center_dist = flat_obj_array([
-        nodes[i] - source_center[i]
+        nodes[i] - center[i]
         for i in range(discr.dim)
         ])
 
@@ -108,7 +128,7 @@ def bump(actx, discr, t=0):
         np.cos(source_omega*t)
         * actx.np.exp(
             -np.dot(center_dist, center_dist)
-            / source_width**2))
+            / width**2))
 
 
 def main():
@@ -137,17 +157,27 @@ def main():
 
     print("%d elements" % mesh.nelements)
 
-    discr = EagerDGDiscretization(actx, mesh, order=order)
+    from meshmode.discretization.poly_element import \
+            QuadratureSimplexGroupFactory, \
+            PolynomialWarpAndBlendGroupFactory
+    discr = DGDiscretizationWithBoundaries(actx, mesh,
+            quad_tag_to_group_factory={
+                QTAG_NONE: PolynomialWarpAndBlendGroupFactory(order),
+                "vel_prod": QuadratureSimplexGroupFactory(3*order),
+                })
+
+    # bounded above by 1
+    c = 0.2 + 0.8*bump(actx, discr, center=np.zeros(3), width=0.5)
 
     fields = flat_obj_array(
-            bump(actx, discr),
+            bump(actx, discr, ),
             [discr.zeros(actx) for i in range(discr.dim)]
             )
 
     vis = make_visualizer(discr, order+3 if dim == 2 else order)
 
     def rhs(t, w):
-        return wave_operator(discr, c=1, w=w)
+        return wave_operator(discr, c=c, w=w)
 
     t = 0
     t_final = 3
@@ -156,10 +186,10 @@ def main():
         fields = rk4_step(fields, t, dt, rhs)
 
         if istep % 10 == 0:
-            print(f"step: {istep} t: {t} L2: {discr.norm(fields[0])} "
-                    f"sol max: {discr.nodal_max('vol', fields[0])}")
-            vis.write_vtk_file("fld-wave-eager-%04d.vtu" % istep,
+            print(istep, t, op.norm(discr, fields[0], p=2))
+            vis.write_vtk_file("fld-wave-eager-var-velocity-%04d.vtu" % istep,
                     [
+                        ("c", c),
                         ("u", fields[0]),
                         ("v", fields[1:]),
                         ])
