+++ examples/wave/wave-op.py
@@ -32,11 +32,10 @@ from meshmode.dof_array import thaw
 
 from meshmode.mesh import BTAG_ALL, BTAG_NONE  # noqa
 
-from grudge.eager import (
-        EagerDGDiscretization, interior_trace_pair, cross_rank_trace_pairs)
+from grudge.discretization import DGDiscretizationWithBoundaries
+import grudge.op as op
 from grudge.shortcuts import make_visualizer
 from grudge.symbolic.primitives import TracePair
-from mpi4py import MPI
 
 
 # {{{ wave equation bits
@@ -45,7 +44,7 @@ def wave_flux(discr, c, w_tpair):
     u = w_tpair[0]
     v = w_tpair[1:]
 
-    normal = thaw(u.int.array_context, discr.normal(w_tpair.dd))
+    normal = thaw(u.int.array_context, op.normal(discr, w_tpair.dd))
 
     flux_weak = flat_obj_array(
             np.dot(v.avg, normal),
@@ -53,40 +52,35 @@ def wave_flux(discr, c, w_tpair):
             )
 
     # upwind
-    v_jump = np.dot(normal, v.ext-v.int)
     flux_weak += flat_obj_array(
             0.5*(u.ext-u.int),
-            0.5*normal*v_jump,
+            0.5*normal*np.dot(normal, v.ext-v.int),
             )
 
-    return discr.project(w_tpair.dd, "all_faces", c*flux_weak)
+    return op.project(discr, w_tpair.dd, "all_faces", c*flux_weak)
 
 
 def wave_operator(discr, c, w):
     u = w[0]
     v = w[1:]
 
-    dir_u = discr.project("vol", BTAG_ALL, u)
-    dir_v = discr.project("vol", BTAG_ALL, v)
+    dir_u = op.project(discr, "vol", BTAG_ALL, u)
+    dir_v = op.project(discr, "vol", BTAG_ALL, v)
     dir_bval = flat_obj_array(dir_u, dir_v)
     dir_bc = flat_obj_array(-dir_u, dir_v)
 
     return (
-            discr.inverse_mass(
+            op.inverse_mass(discr,
                 flat_obj_array(
-                    -c*discr.weak_div(v),
-                    -c*discr.weak_grad(u)
+                    -c*op.weak_div(discr, v),
+                    -c*op.weak_grad(discr, u)
                     )
                 +  # noqa: W504
-                discr.face_mass(
-                    wave_flux(discr, c=c, w_tpair=interior_trace_pair(discr, w))
+                op.face_mass(discr,
+                    wave_flux(discr, c=c, w_tpair=op.interior_trace_pair(discr, w))
                     + wave_flux(discr, c=c, w_tpair=TracePair(
                         BTAG_ALL, interior=dir_bval, exterior=dir_bc))
-                    + sum(
-                        wave_flux(discr, c=c, w_tpair=tpair)
-                        for tpair in cross_rank_trace_pairs(discr, w))
-                    )
-                )
+                    ))
                 )
 
 # }}}
@@ -105,7 +99,7 @@ def bump(actx, discr, t=0):
     source_width = 0.05
     source_omega = 3
 
-    nodes = thaw(actx, discr.nodes())
+    nodes = thaw(actx, op.nodes(discr))
     center_dist = flat_obj_array([
         nodes[i] - source_center[i]
         for i in range(discr.dim)
@@ -123,37 +117,15 @@ def main():
     queue = cl.CommandQueue(cl_ctx)
     actx = PyOpenCLArrayContext(queue)
 
-    comm = MPI.COMM_WORLD
-    num_parts = comm.Get_size()
-
-    from meshmode.distributed import MPIMeshDistributor, get_partition_by_pymetis
-    mesh_dist = MPIMeshDistributor(comm)
-
     dim = 2
     nel_1d = 16
-
-    if mesh_dist.is_mananger_rank():
-        from meshmode.mesh.generation import generate_regular_rect_mesh
-        mesh = generate_regular_rect_mesh(
-                a=(-0.5,)*dim,
-                b=(0.5,)*dim,
-                n=(nel_1d,)*dim)
-
-        print("%d elements" % mesh.nelements)
-
-        part_per_element = get_partition_by_pymetis(mesh, num_parts)
-
-        local_mesh = mesh_dist.send_mesh_parts(mesh, part_per_element, num_parts)
-
-        del mesh
-
-    else:
-        local_mesh = mesh_dist.receive_mesh_part()
+    from meshmode.mesh.generation import generate_regular_rect_mesh
+    mesh = generate_regular_rect_mesh(
+            a=(-0.5,)*dim,
+            b=(0.5,)*dim,
+            n=(nel_1d,)*dim)
 
     order = 3
-
-    discr = EagerDGDiscretization(actx, local_mesh, order=order,
-                    mpi_communicator=comm)
 
     if dim == 2:
         # no deep meaning here, just a fudge factor
@@ -163,6 +135,10 @@ def main():
         dt = 0.45/(nel_1d*order**2)
     else:
         raise ValueError("don't have a stable time step guesstimate")
+
+    print("%d elements" % mesh.nelements)
+
+    discr = DGDiscretizationWithBoundaries(actx, mesh, order=order)
 
     fields = flat_obj_array(
             bump(actx, discr),
@@ -181,10 +157,9 @@ def main():
         fields = rk4_step(fields, t, dt, rhs)
 
         if istep % 10 == 0:
-            print(istep, t, discr.norm(fields[0]))
-            vis.write_parallel_vtk_file(
-                    comm,
-                    f"fld-wave-eager-mpi-{{rank:03d}}-{istep:04d}.vtu",
+            print(f"step: {istep} t: {t} L2: {op.norm(discr, fields[0], 2)} "
+                    f"sol max: {op.nodal_max(discr, 'vol', fields[0])}")
+            vis.write_vtk_file("fld-wave-eager-%04d.vtu" % istep,
                     [
                         ("u", fields[0]),
                         ("v", fields[1:]),
